✨ index.js作成
