+++ src/Sass/boundAnimation.scss
