+++ src/css/boundAnimation.css
