+++ src/js/components/IntersectionObserver-polyfill.js
