+++ src/js/components/cursor.js
